Add some spaces to clarify meaning
Although not a bug, I chose to address this interpretation difference
first.  We're learning so we migth as well learn.

Lisp is read as PAREN_OPEN functionNameOrBody any arguments follow here
CLOSE_PAREN.

The publisher would be written in javascript as:

    defineResource( "publisher", [], {
      class: sPrefix("ext:Publisher"),
      properties: [["name", "string", sPrefix("schema:name")]],
      hasMany: [["book", "via", sPrefix("ext:publisher")]],
      resourceBase: sUrl("https://mu.semte.ch/examples/publishers/"),
      onPath: "publishers"
    })

Notice that lisp writes `(define-resource ...)` and javascript writes
`defineResource(...)`.

A json representation would be slightly more verbose.

The parens `()` after publisher are a placeholder to specify
inheritance (you could state that a book is a creative-work).